tool_use/email_agent/requirements.txt
# Email Agent Specific Dependencies
# Install from project root: pip install -r tool_use/email_agent/requirements.txt

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.0.0
pydantic[email]  # Email validation support
email-validator

# Environment & HTTP
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Interface (uses parent tool_use package)
aisuite>=0.1.11

# Markdown rendering for LLM service
markdown>=3.5.0
